backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
google-auth==2.27.0
stripe==7.12.0
celery==5.3.6
django-celery-beat==2.6.0
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
gunicorn==21.2.0
uvicorn==0.27.1
